chore(catalog): Add schema/types
