+++ internal/provider/catalog/catalog_types.go
@@ -35,14 +35,3 @@ type catalogResourceModel struct {
 	DeleteForce     types.Bool   `tfsdk:"delete_force"`
 	DeleteRecursive types.Bool   `tfsdk:"delete_recursive"`
 }
-
-type vAppTemplateDataSourceModel struct {
-	ID           types.String `tfsdk:"id"`
-	TemplateName types.String `tfsdk:"template_name"`
-	TemplateID   types.String `tfsdk:"template_id"`
-	CatalogID    types.String `tfsdk:"catalog_id"`
-	CatalogName  types.String `tfsdk:"catalog_name"`
-	Description  types.String `tfsdk:"description"`
-	CreatedAt    types.String `tfsdk:"created_at"`
-	VMNames      types.List   `tfsdk:"vm_names"`
-}
