+++ internal/provider/catalog/media_schema.go
@@ -1,12 +1,8 @@
 package catalog
 
 import (
-	"context"
-
-	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 
 	schemaD "github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	schemaR "github.com/hashicorp/terraform-plugin-framework/resource/schema"
@@ -18,64 +14,6 @@ import (
 
 	superschema "github.com/FrangipaneTeam/terraform-plugin-framework-superschema"
 )
-
-type catalogMediaDataSourceModel struct {
-	ID             types.String `tfsdk:"id"`
-	Name           types.String `tfsdk:"name"`
-	CatalogID      types.String `tfsdk:"catalog_id"`
-	CatalogName    types.String `tfsdk:"catalog_name"`
-	Description    types.String `tfsdk:"description"`
-	IsISO          types.Bool   `tfsdk:"is_iso"`
-	OwnerName      types.String `tfsdk:"owner_name"`
-	IsPublished    types.Bool   `tfsdk:"is_published"`
-	CreatedAt      types.String `tfsdk:"created_at"`
-	Size           types.Int64  `tfsdk:"size"`
-	Status         types.String `tfsdk:"status"`
-	StorageProfile types.String `tfsdk:"storage_profile"`
-}
-
-func catalogMediaDataSourceModelType() map[string]attr.Type {
-	return map[string]attr.Type{
-		"id":              types.StringType,
-		"name":            types.StringType,
-		"catalog_id":      types.StringType,
-		"catalog_name":    types.StringType,
-		"description":     types.StringType,
-		"is_iso":          types.BoolType,
-		"owner_name":      types.StringType,
-		"is_published":    types.BoolType,
-		"created_at":      types.StringType,
-		"size":            types.Int64Type,
-		"status":          types.StringType,
-		"storage_profile": types.StringType,
-	}
-}
-
-func mediaDatasourceAttributes() map[string]schemaD.Attribute {
-	attr := mediaSchema().GetDataSource(context.Background()).Attributes
-
-	for k, v := range attr {
-		switch val := v.(type) {
-		case schemaD.StringAttribute:
-			val.Computed = true
-			val.Optional = false
-			val.Required = false
-			attr[k] = val
-		case schemaD.BoolAttribute:
-			val.Computed = true
-			val.Optional = false
-			val.Required = false
-			attr[k] = val
-		case schemaD.Int64Attribute:
-			val.Computed = true
-			val.Optional = false
-			val.Required = false
-			attr[k] = val
-		}
-	}
-
-	return attr
-}
 
 func mediaSchema() superschema.Schema {
 	return superschema.Schema{
