+++ internal/provider/catalog/media_types.go
@@ -1,7 +1,6 @@
 package catalog
 
 import (
-	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
 
@@ -18,21 +17,4 @@ type catalogMediaDataSourceModel struct {
 	Size           types.Int64  `tfsdk:"size"`
 	Status         types.String `tfsdk:"status"`
 	StorageProfile types.String `tfsdk:"storage_profile"`
-}
-
-func catalogMediaDataSourceModelType() map[string]attr.Type {
-	return map[string]attr.Type{
-		"id":              types.StringType,
-		"name":            types.StringType,
-		"catalog_id":      types.StringType,
-		"catalog_name":    types.StringType,
-		"description":     types.StringType,
-		"is_iso":          types.BoolType,
-		"owner_name":      types.StringType,
-		"is_published":    types.BoolType,
-		"created_at":      types.StringType,
-		"size":            types.Int64Type,
-		"status":          types.StringType,
-		"storage_profile": types.StringType,
-	}
 }
